pyproject.toml
[tool.poetry]
name = "ghost_post_backend"
version = "0.1.0"
description = ""
authors = ["Kyle Negley"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-cors-headers = "^3.5.0"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
